3
3

+++ dragonpoop_alpha/dragonpoop/gfx/model/model_loader/model_loader_ms3d/model_loader_ms3d_animation.h
@@ -18,7 +18,7 @@ namespace dragonpoop
     
     struct ms3d_model_frame
     {
-        float t;
+        unsigned int t;
         dpid id, afid;
     };
